Revert "config/start-angular-project"
