+++ backend-module/src/main/java/com/example/backend/BackendApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class BackendApplication {
-  public static void main(String[] args) {
-    SpringApplication.run(BackendApplication.class, args);
-  }
+    public static void main(String[] args) {
+        SpringApplication.run(BackendApplication.class, args);
+    }
 }
