--- backend-module/src/main/resources/wildfly-config.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-    <authentication-client xmlns="urn:elytron:client:1.2">
-        <authentication-rules>
-            <rule use-configuration="default"/>
-        </authentication-rules>
-        <authentication-configurations>
-            <configuration name="default">
-                <sasl-mechanism-selector selector="DIGEST-MD5"/>
-                <credentials>
-                    <clear-password password="app123"/>
-                </credentials>
-                <set-user-name name="appuser"/>
-            </configuration>
-        </authentication-configurations>
-    </authentication-client>
-
-    <jboss-ejb-client xmlns="urn:jboss:wildfly-client-ejb:3.2">
-        <connections>
-            <connection uri="remote+http://localhost:8083"/>
-        </connections>
-    </jboss-ejb-client>
-</configuration>
