--- ejb-module/README.md
@@ -1,3 +1,0 @@
-# Frontend Angular
-
-Starter Angular project deve ser criado pelo candidato.
