+++ ejb-module/src/main/java/com/example/ejb/Beneficio.java
@@ -1,0 +1,27 @@
+package com.example.ejb;
+
+import java.math.BigDecimal;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Table(name = "beneficio")
+@Getter
+public class Beneficio {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Setter
+  @Column(nullable = false)
+  private BigDecimal valor;
+
+}
