--- ejb-module/src/main/resources/META-INF/beans.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans bean-discovery-mode="all"
-       xmlns="https://jakarta.ee/xml/ns/jakartaee"/>
