--- frontend/src/app/app.config.ts
@@ -1,13 +1,0 @@
-import { ApplicationConfig } from '@angular/core';
-import { provideHttpClient } from '@angular/common/http';
-import { provideRouter } from '@angular/router';
-import { routes } from './app.routes';
-import { BENEFICIO_PROVIDERS } from './domains/beneficio/beneficio.providers';
-
-export const appConfig: ApplicationConfig = {
-  providers: [
-    provideRouter(routes),
-    provideHttpClient(),
-    ...BENEFICIO_PROVIDERS,
-  ],
-};
