--- frontend/src/app/app.html
@@ -1,2 +1,0 @@
-
-<router-outlet />
