--- frontend/src/app/domains/beneficio/beneficio-list.component.html
@@ -1,36 +1,0 @@
-<div>
-  <h1>Benefícios</h1>
-  <button (click)="router.navigate(['/beneficios/novo'])">+ Novo</button>
-
-  @if (loading()) {
-    <p>Carregando...</p>
-  } @else if (error()) {
-    <p>{{ error() }}</p>
-  } @else {
-    <table>
-      <thead>
-      <tr>
-        <th>Nome</th>
-        <th>Descrição</th>
-        <th>Valor</th>
-        <th>Ativo</th>
-        <th>Ações</th>
-      </tr>
-      </thead>
-      <tbody>
-        @for (b of beneficios(); track b.id) {
-          <tr>
-            <td>{{ b.nome }}</td>
-            <td>{{ b.descricao }}</td>
-            <td>{{ b.valor | currency:'BRL' }}</td>
-            <td>{{ b.ativo ? 'Sim' : 'Não' }}</td>
-            <td>
-              <button (click)="router.navigate(['/beneficios', b.id, 'editar'])">Editar</button>
-              <button (click)="delete(b.id!)">Excluir</button>
-            </td>
-          </tr>
-        }
-      </tbody>
-    </table>
-  }
-</div>
