--- frontend/src/app/domains/beneficio/beneficio.model.ts
@@ -1,7 +1,0 @@
-export interface Beneficio {
-  id: number | null;
-  nome: string;
-  descricao: string;
-  ativo: boolean;
-  valor: number;
-}
